requirements.txt
timm
gradio
pillow
matplotlib==3.5.3
monai
numpy==1.21.6
pandas==1.3.5
seaborn==0.11.2
session_info==1.0.0
torch==1.11.0
torchvision==0.12.0
tqdm
